Add initial version of peer binding.
This is an initial version because as I explored, I realised that
things could be optimised. For example, only one ssh_bind object
is needed, and that can be initialised with duplex_peer. Also,
by managing fds on our own, we can take advantage of select() and
make lives easier when a client calls accept() on the peer.

So this needs a bit of rethinking. There are also some new FIXMEs
where there may be fd leaks (grep LEAK), and I commented out most
of goduplex's tests right now while we figure out the problem with
global requests.